+++ prototypes.h
@@ -13,6 +13,7 @@
 #include <stdlib.h>
 #include <cmath>
 #include <cstdlib>
+#include <vector>
 
 using namespace std;
 // defining the value of PI.
@@ -69,7 +70,8 @@ double der_wavf_func (double t[],double t_der[]);
 // -----------------------------------------------------------------------------------------------
 //
 
-double mesh_gen_o (float area_x, float area_y, int n_x, int n_y, float X[], float Y[]);
+
+int mesh_gen_o (float area_x, float area_y, int n_x, int n_y, float X[], float Y[]);
 
 
 //
